# openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BhashaAI Backend",
    "description": "\n## BhashaAI - Generative AI for Education & Gujarati Language Technologies\n\nA comprehensive API for educational AI services supporting:\n\n- **Question Paper Generation** - AI-powered question paper creation from educational content\n- **Assignment Help** - Socratic tutoring and step-by-step solutions\n- **Teaching Tools** - Mind maps, slides, and study materials generation\n- **Paper Checking** - OCR-based automated evaluation with Gujarati support\n- **Gujarati Learning** - Vocabulary, grammar, and gamified learning\n- **Audio Services** - STT/TTS for Gujarati language\n\n### Authentication\nAll protected endpoints require JWT Bearer token authentication.\n\n### Bilingual Support\nResponses include both English and Gujarati (ગુજરાતી) messages.\n        ",
    "contact": {
      "name": "BhashaAI Team",
      "email": "support@bhashaai.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents/upload": {
      "post": {
        "tags": [
          "Documents",
          "Documents"
        ],
        "summary": "Upload Document",
        "description": "Upload a PDF, DOCX, or TXT document for question generation.",
        "operationId": "upload_document_api_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/documents": {
      "get": {
        "tags": [
          "Documents",
          "Documents"
        ],
        "summary": "List Documents",
        "description": "List all documents for the current user.",
        "operationId": "list_documents_api_v1_documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "Documents",
          "Documents"
        ],
        "summary": "Get Document",
        "description": "Get document details by ID.",
        "operationId": "get_document_api_v1_documents__document_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Documents",
          "Documents"
        ],
        "summary": "Delete Document",
        "description": "Delete a document (soft delete).",
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_id}/download": {
      "get": {
        "tags": [
          "Documents",
          "Documents"
        ],
        "summary": "Download Document",
        "description": "Download the document file.",
        "operationId": "download_document_api_v1_documents__document_id__download_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful"
          },
          "data": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response data payload"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Human-readable message (English)"
          },
          "message_gu": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Gu",
            "description": "Human-readable message (Gujarati / ગુજરાતી)"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "APIResponse",
        "description": "Standard API response envelope.\n\nAll API responses follow this structure for consistency.\n\nAttributes:\n    success: Whether the operation was successful\n    data: Response payload (type varies by endpoint)\n    message: Human-readable message in English\n    message_gu: Human-readable message in Gujarati\n\nExample:\n    ```json\n    {\n        \"success\": true,\n        \"data\": {\"id\": \"uuid-here\", \"name\": \"Example\"},\n        \"message\": \"Operation successful\",\n        \"message_gu\": \"કામગીરી સફળ\"\n    }\n    ```"
      },
...
      "Body_upload_document_api_v1_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "grade_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Level"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_documents_upload_post"
      },
...
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "institution_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "file_url": {
            "type": "string",
            "title": "File Url"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "text_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Content"
          },
          "extra_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra Metadata",
            "default": {}
          },
          "processing_status": {
            "type": "string",
            "title": "Processing Status"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "grade_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Level"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "filename",
          "file_url",
          "file_type",
          "file_size",
          "mime_type",
          "processing_status",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentResponse",
        "description": "Schema for document response."
      },
      "DocumentUploadResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Document uploaded successfully"
          },
          "message_gu": {
            "type": "string",
            "title": "Message Gu",
            "default": "ડોક્યુમેન્ટ સફળતાપૂર્વક અપલોડ થયું"
          },
          "data": {
            "$ref": "#/components/schemas/DocumentResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DocumentUploadResponse",
        "description": "Schema for document upload response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /api/v1/auth/login"
      },
      "RefreshToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Refresh-Token",
        "description": "JWT refresh token for obtaining new access tokens"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.bhashaai.com",
      "description": "Production server"
    }
  ]
}